+++ rust/pact_ffi/src/verifier/handle.rs
@@ -4,14 +4,16 @@ use std::sync::Arc;
 
 use log::debug;
 
-use pact_verifier::{FilterInfo, NullRequestFilterExecutor, PactSource, ProviderInfo, VerificationOptions, verify_provider_async};
+use pact_models::prelude::HttpAuth;
+use pact_verifier::{FilterInfo, NullRequestFilterExecutor, PactSource, ProviderInfo, VerificationOptions, verify_provider_async, ConsumerVersionSelector};
 use pact_verifier::callback_executors::HttpRequestProviderStateExecutor;
 
 #[derive(Debug, Clone)]
 /// Wraps a Pact verifier
 pub struct VerifierHandle {
   provider: ProviderInfo,
-  sources: Vec<PactSource>
+  sources: Vec<PactSource>,
+  state_change: Arc<HttpRequestProviderStateExecutor>
 }
 
 impl VerifierHandle {
@@ -19,7 +21,8 @@ impl VerifierHandle {
   pub fn new() -> VerifierHandle {
     VerifierHandle {
       provider: ProviderInfo::default(),
-      sources: Vec::new()
+      sources: Vec::new(),
+      state_change: Arc::new(HttpRequestProviderStateExecutor::default())
     }
   }
 
@@ -46,6 +49,75 @@ impl VerifierHandle {
     self.sources.push(PactSource::File(file.to_string()));
   }
 
+  /// Add a directory source to be verified. This will verify all pact files in the directory.
+  pub fn add_directory_source(&mut self, dir: &str) {
+    self.sources.push(PactSource::Dir(dir.to_string()));
+  }
+
+  /// Add a URL source to be verified. This will fetch the pact file from the URL. If a username
+  /// and password is given, then basic authentication will be used when fetching the pact file.
+  /// If a token is provided, then bearer token authentication will be used.
+  pub fn add_url_source(&mut self, url: &str, auth: &HttpAuth) {
+    if !auth.is_none() {
+      self.sources.push(PactSource::URL(url.to_string(), Some(auth.clone())));
+    } else {
+      self.sources.push(PactSource::URL(url.to_string(), None));
+    }
+  }
+
+  /// Add a Pact broker source to be verified. This will fetch all the pact files from the broker
+  /// that match the provider name. If a username
+  /// and password is given, the basic authentication will be used when fetching the pact file.
+  /// If a token is provided, then bearer token authentication will be used.
+  pub fn add_pact_broker_source(
+    &mut self,
+    url: &str,
+    provider_name: &str,
+    enable_pending: bool,
+    include_wip_pacts_since: Option<String>,
+    provider_tags: Vec<String>,
+    selectors: Vec<ConsumerVersionSelector>,
+    auth: &HttpAuth
+  ) {
+    if !auth.is_none() {
+      self.sources.push(PactSource::BrokerWithDynamicConfiguration {
+        provider_name: provider_name.to_string(),
+        broker_url: url.to_string(),
+        enable_pending,
+        include_wip_pacts_since,
+        provider_tags,
+        selectors,
+        auth: Some(auth.clone()),
+        links: vec![]
+      });
+    } else {
+      self.sources.push(PactSource::BrokerWithDynamicConfiguration {
+        provider_name: provider_name.to_string(),
+        broker_url: url.to_string(),
+        enable_pending,
+        include_wip_pacts_since,
+        provider_tags,
+        selectors,
+        auth: None,
+        links: vec![]
+      });
+    }
+  }
+
+  /// Update the provider state
+  pub fn update_provider_state(
+    &mut self,
+    state_change_url: &str,
+    state_change_teardown: bool,
+    state_change_body: bool
+  ) {
+    self.state_change = Arc::new(HttpRequestProviderStateExecutor {
+      state_change_url: Some(state_change_url.to_string()),
+      state_change_teardown,
+      state_change_body
+    })
+  }
+
   /// Execute the verifier
   ///
   /// This will return an integer value based on the status of the verification:
@@ -54,11 +126,6 @@ impl VerifierHandle {
   /// * 2 - failed to run the verification
   pub fn execute(&self) -> i32 {
     let filter = FilterInfo::None;
-    let provider_state_executor = Arc::new(HttpRequestProviderStateExecutor {
-      state_change_url: None,
-      state_change_teardown: false,
-      state_change_body: false
-    });
 
     let options = VerificationOptions {
       request_filter: None::<Arc<NullRequestFilterExecutor>>,
@@ -77,7 +144,7 @@ impl VerifierHandle {
         filter,
         vec![],
         options,
-        &provider_state_executor
+        &self.state_change.clone()
       ).await
     })
       .map(|result| if result { 0 } else { 2 })