+++ rust/pact_ffi/src/verifier/mod.rs
@@ -3,16 +3,21 @@
 
 #![warn(missing_docs)]
 
-use std::ffi::CStr;
+use std::ffi::{CStr, CString, OsStr, OsString};
 use std::panic::catch_unwind;
 
 use anyhow::Context;
-use libc::{c_char, c_int, c_ushort, EXIT_FAILURE};
+use libc::{c_char, c_int, c_uchar, c_ushort, EXIT_FAILURE};
 use log::*;
+use std::env;
+
+use pact_models::prelude::HttpAuth;
 
 use crate::{as_mut, ffi_fn, safe_str};
 use crate::util::*;
 use crate::util::string::if_null;
+use serde::{Serialize, Deserialize};
+use clap::ArgSettings;
 
 mod args;
 pub mod verifier;
@@ -119,6 +124,30 @@ ffi_fn! {
 }
 
 ffi_fn! {
+    /// Set the provider state for the Pact verifier.
+    ///
+    /// `teardown` is a boolean value. Set it to greater than zero to turn the option on.
+    /// `body` is a boolean value. Set it to greater than zero to turn the option on.
+    ///
+    /// # Safety
+    ///
+    /// All string fields must contain valid UTF-8. Invalid UTF-8
+    /// will be replaced with U+FFFD REPLACEMENT CHARACTER.
+    ///
+    fn pactffi_verifier_update_provider_state(
+      handle: *mut handle::VerifierHandle,
+      url: *const c_char,
+      teardown: c_uchar,
+      body: c_uchar
+    ) {
+      let handle = as_mut!(handle);
+      let url = safe_str!(url);
+
+      handle.update_provider_state(url, teardown > 0, body > 0);
+    }
+}
+
+ffi_fn! {
     /// Adds a Pact file as a source to verify.
     ///
     /// # Safety
@@ -138,6 +167,192 @@ ffi_fn! {
 }
 
 ffi_fn! {
+    /// Adds a Pact directory as a source to verify. All pacts from the directory that match the
+    /// provider name will be verified.
+    ///
+    /// # Safety
+    ///
+    /// All string fields must contain valid UTF-8. Invalid UTF-8
+    /// will be replaced with U+FFFD REPLACEMENT CHARACTER.
+    ///
+    fn pactffi_verifier_add_directory_source(
+      handle: *mut handle::VerifierHandle,
+      directory: *const c_char
+    ) {
+      let handle = as_mut!(handle);
+      let directory = safe_str!(directory);
+
+      handle.add_directory_source(directory);
+    }
+}
+
+ffi_fn! {
+    /// Adds a URL as a source to verify. The Pact file will be fetched from the URL.
+    ///
+    /// If a username and password is given, then basic authentication will be used when fetching
+    /// the pact file. If a token is provided, then bearer token authentication will be used.
+    ///
+    /// # Safety
+    ///
+    /// All string fields must contain valid UTF-8. Invalid UTF-8
+    /// will be replaced with U+FFFD REPLACEMENT CHARACTER.
+    ///
+    fn pactffi_verifier_url_source(
+      handle: *mut handle::VerifierHandle,
+      url: *const c_char,
+      username: *const c_char,
+      password: *const c_char,
+      token: *const c_char
+    ) {
+      let handle = as_mut!(handle);
+      let url = safe_str!(url);
+      let username = if_null(username, "");
+      let password = if_null(password, "");
+      let token = if_null(token, "");
+
+      let auth = if !username.is_empty() {
+        if !password.is_empty() {
+          HttpAuth::User(username, Some(password))
+        } else {
+          HttpAuth::User(username, None)
+        }
+      } else if !token.is_empty() {
+        HttpAuth::Token(token)
+      } else {
+        HttpAuth::None
+      };
+
+      handle.add_url_source(url, &auth);
+    }
+}
+
+ffi_fn! {
+    /// Adds a Pact broker as a source to verify. This will fetch all the pact files from the broker
+    /// that match the provider name.
+    ///
+    /// If a username and password is given, then basic authentication will be used when fetching
+    /// the pact file. If a token is provided, then bearer token authentication will be used.
+    ///
+    /// # Safety
+    ///
+    /// All string fields must contain valid UTF-8. Invalid UTF-8
+    /// will be replaced with U+FFFD REPLACEMENT CHARACTER.
+    ///
+    fn pactffi_verifier_broker_source(
+      handle: *mut handle::VerifierHandle,
+      url: *const c_char,
+      provider_name: *const c_char,
+      username: *const c_char,
+      password: *const c_char,
+      token: *const c_char
+    ) {
+      let handle = as_mut!(handle);
+      let url = safe_str!(url);
+      let provider_name = safe_str!(provider_name);
+      let username = if_null(username, "");
+      let password = if_null(password, "");
+      let token = if_null(token, "");
+
+      let auth = if !username.is_empty() {
+        if !password.is_empty() {
+          HttpAuth::User(username, Some(password))
+        } else {
+          HttpAuth::User(username, None)
+        }
+      } else if !token.is_empty() {
+        HttpAuth::Token(token)
+      } else {
+        HttpAuth::None
+      };
+
+      handle.add_pact_broker_source(url, provider_name, false, None, vec![], vec![], &auth);
+    }
+}
+
+ffi_fn! {
+    /// Adds a Pact broker as a source to verify. This will fetch all the pact files from the broker
+    /// that match the provider name and the consumer version selectors
+    /// (See `https://docs.pact.io/pact_broker/advanced_topics/consumer_version_selectors/`).
+    ///
+    /// The consumer version selectors must be passed in in JSON format.
+    ///
+    /// `enable_pending` is a boolean value. Set it to greater than zero to turn the option on.
+    ///
+    /// If the `include_wip_pacts_since` option is provided, it needs to be a date formatted in
+    /// ISO format (YYYY-MM-DD).
+    ///
+    /// If a username and password is given, then basic authentication will be used when fetching
+    /// the pact file. If a token is provided, then bearer token authentication will be used.
+    ///
+    /// # Safety
+    ///
+    /// All string fields must contain valid UTF-8. Invalid UTF-8
+    /// will be replaced with U+FFFD REPLACEMENT CHARACTER.
+    ///
+    fn pactffi_verifier_broker_source_with_selectors(
+      handle: *mut handle::VerifierHandle,
+      url: *const c_char,
+      provider_name: *const c_char,
+      username: *const c_char,
+      password: *const c_char,
+      token: *const c_char,
+      enable_pending: c_uchar,
+      include_wip_pacts_since: *const c_char,
+      provider_tags: *const *const c_char,
+      provider_tags_len: c_ushort
+    ) {
+      let handle = as_mut!(handle);
+      let url = safe_str!(url);
+      let provider_name = safe_str!(provider_name);
+      let username = if_null(username, "");
+      let password = if_null(password, "");
+      let token = if_null(token, "");
+      let wip_pacts = if_null(include_wip_pacts_since, "");
+
+      let auth = if !username.is_empty() {
+        if !password.is_empty() {
+          HttpAuth::User(username, Some(password))
+        } else {
+          HttpAuth::User(username, None)
+        }
+      } else if !token.is_empty() {
+        HttpAuth::Token(token)
+      } else {
+        HttpAuth::None
+      };
+
+      let wip = if !wip_pacts.is_empty() {
+        Some(wip_pacts)
+      } else {
+        None
+      };
+
+      let tags = if !provider_tags.is_null() && provider_tags_len > 0 {
+        let mut tags = Vec::with_capacity(provider_tags_len as usize);
+        for index in 0..(provider_tags_len - 1) {
+          let tag_ptr: * const c_char = unsafe { *(provider_tags.offset(index as isize)) };
+          tags.push(safe_str!(tag_ptr).to_string());
+        }
+        tags
+      } else {
+        vec![]
+      };
+
+    // let selectors = if matches.is_present("consumer-version-selectors") {
+    // matches.values_of("consumer-version-selectors")
+    // .map_or_else(Vec::new, |s| json_to_selectors(s.collect::<Vec<_>>()))
+    // } else if matches.is_present("consumer-version-tags") {
+    // matches.values_of("consumer-version-tags")
+    // .map_or_else(Vec::new, |tags| consumer_tags_to_selectors(tags.collect::<Vec<_>>()))
+    // } else {
+    // vec![]
+    // };
+
+      handle.add_pact_broker_source(url, provider_name, enable_pending > 0, wip, tags, vec![], &auth);
+    }
+}
+
+ffi_fn! {
     /// Runs the verification.
     ///
     /// # Error Handling
@@ -150,4 +365,168 @@ ffi_fn! {
     } {
       EXIT_FAILURE
     }
+}
+
+/// Contain the various attributes of an argument given to the verifier
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Argument {
+    long: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    short: Option<String>,
+    help: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    possible_values: Option<Vec<String>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    default_value: Option<String>,
+    multiple: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    env: Option<String>,
+}
+
+/// Contain the lists of the two types of argument: options and flags
+#[derive(Debug, Serialize, Deserialize)]
+pub struct OptionsFlags {
+    /// Arguments which require a parameter, such as loglevel
+    pub options: Vec<Argument>,
+    /// Arguments which are a bool, such as publish
+    pub flags: Vec<Argument>
+}
+
+/// External interface to retrieve the options and arguments available when calling the CLI interface,
+/// returning them as a JSON string.
+///
+/// The purpose is to then be able to use in other languages which wrap the FFI library, to implement
+/// the same CLI functionality automatically without manual maintenance of arguments, help descriptions
+/// etc.
+///
+/// # Example structure
+/// ```json
+/// {
+///   "options": [
+///     {
+///       "long": "scheme",
+///       "help": "Provider URI scheme (defaults to http)",
+///       "possible_values": [
+///         "http",
+///         "https"
+///       ],
+///       "default_value": "http"
+///       "multiple": false,
+///     },
+///     {
+///       "long": "file",
+///       "short": "f",
+///       "help": "Pact file to verify (can be repeated)",
+///       "multiple": true
+///     },
+///     {
+///       "long": "user",
+///       "help": "Username to use when fetching pacts from URLS",
+///       "multiple": false,
+///       "env": "PACT_BROKER_USERNAME"
+///     }
+///   ],
+///   "flags": [
+///     {
+///       "long": "disable-ssl-verification",
+///       "help": "Disables validation of SSL certificates",
+///       "multiple": false
+///     }
+///   ]
+/// }
+/// ```
+///
+/// # Safety
+///
+/// Exported functions are inherently unsafe.
+#[no_mangle]
+pub extern "C" fn pactffi_verifier_cli_args() -> *const c_char {
+    let args: Vec<String> = env::args().collect();
+    let program = args[0].clone();
+    let app = args::setup_app(program, clap::crate_version!());
+
+    // Iterate through the args, extracting info from each to then add to a Vector of args
+    let mut options: Vec<Argument> = Vec::new();
+    let mut flags: Vec<Argument> = Vec::new();
+
+    for opt in app.p.opts.iter() {
+        let arg = parse_argument(opt.s.long, opt.s.short, opt.b.help, opt.v.possible_vals.clone(),opt.v.default_val, opt.b.settings.is_set(ArgSettings::Multiple), opt.v.env.clone());
+        options.push(arg);
+    }
+
+    for opt in app.p.flags.iter() {
+        let arg = parse_argument(opt.s.long, opt.s.short, opt.b.help, None, None,opt.b.settings.is_set(ArgSettings::Multiple), None);
+        flags.push(arg);
+    }
+
+
+    let opts_flags = OptionsFlags { options:options, flags: flags};
+    let json = serde_json::to_string(&opts_flags).unwrap();
+    let c_str = CString::new(json).unwrap();
+    c_str.into_raw() as *const c_char
+}
+
+fn parse_argument(long: Option<&str>, short: Option<char>, help: Option<&str>, possible_values: Option<Vec<&str>>, default_value: Option<&OsStr>, multiple: bool, env: Option<(&OsStr, Option<OsString>)>) -> Argument {
+    let mut arg = Argument { short: None, long: None, help: None, possible_values: None, default_value: None, multiple: Some(false), env: None };
+
+    // Long
+    match long {
+        None => {}
+        Some(val) => {
+            arg.long = Some(val.to_string());
+        }
+    }
+
+    // Short
+    match short {
+        None => {}
+        Some(val) => {
+            arg.short = Some(val.to_string());
+        }
+    }
+
+    // Help
+    match help {
+        None => {}
+        Some(val) => {
+            arg.help = Some(val.to_string());
+        }
+    }
+
+    // Possible values
+    match possible_values {
+        None => {}
+        Some(val) => {
+            let mut possible_vals: Vec<String> = Vec::new();
+            for possible_val in val.iter() {
+                possible_vals.push(possible_val.to_string())
+            }
+            arg.possible_values = Some(possible_vals);
+        }
+    }
+
+    // Default value
+    match default_value {
+        None => {}
+        Some(val) =>
+            {
+                arg.default_value = Some(val.to_os_string().into_string().unwrap());
+            }
+    }
+
+    // Multiple
+    if multiple {
+        arg.multiple = Some(true);
+    }
+
+    // Env
+    match env {
+        None => {}
+        Some(val) =>
+            {
+                arg.env = Some(val.0.to_os_string().into_string().unwrap());
+            }
+    }
+
+    arg
 }