+++ rust/pact_models/src/message_pact.rs
@@ -17,9 +17,9 @@ use maplit::*;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
-use crate::{Consumer, http_utils, PactSpecification, Provider};
-use crate::file_utils::with_read_lock;
-use crate::http_utils::HttpAuth;
+use crate::{Consumer, PactSpecification, Provider};
+#[cfg(not(target_family = "wasm"))] use crate::file_utils::with_read_lock;
+#[cfg(not(target_family = "wasm"))] use crate::http_utils::{self, HttpAuth};
 use crate::interaction::Interaction;
 use crate::message::Message;
 use crate::pact::{determine_spec_version, Pact, parse_meta_data, ReadWritePact};
@@ -32,7 +32,7 @@ use crate::verify_json::{json_type_of, PactFileVerificationResult, PactJsonVerif
 /// Struct that represents a pact between the consumer and provider of a service.
 /// It contains a list of Messages instead of Interactions, but is otherwise
 /// identical to `struct Pact`.
-#[derive(Debug, Clone, Deserialize, Serialize)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub struct MessagePact {
     /// Consumer side of the pact
     pub consumer: Consumer,
@@ -219,6 +219,7 @@ impl MessagePact {
 
     /// Reads the pact file from a URL and parses the resulting JSON
     /// into a `MessagePact` struct
+    #[cfg(not(target_family = "wasm"))]
     pub fn from_url(url: &String, auth: &Option<HttpAuth>) -> anyhow::Result<MessagePact> {
         let (url, json) = http_utils::fetch_json_from_url(url, auth)?;
         MessagePact::from_json(&url, &json)
@@ -227,6 +228,7 @@ impl MessagePact {
     /// Writes this pact out to the provided file path.
     /// All directories in the path will automatically created.
     /// If there is already a file at the path, it will be overwritten.
+    #[cfg(not(target_family = "wasm"))]
     pub fn overwrite_pact(
         &self,
         path: &Path,
@@ -273,6 +275,7 @@ impl MessagePact {
 }
 
 impl ReadWritePact for MessagePact {
+  #[cfg(not(target_family = "wasm"))]
   fn read_pact(path: &Path) -> anyhow::Result<MessagePact> {
     with_read_lock(path, 3, &mut |f| {
       let pact_json: Value = serde_json::from_reader(f)?;